main: Add more registers
Looks like this is where the experiment ends. Stack and data memory
are stating to clash resulting in unreliable operation. There
are not a whole lot of other controllers out there that satisfy
the following requirements:

* > 2 kB SRAM
* > 16 kB Flash
* > 512 B EEPROM
* VCC range from 2.7 to 4.3 V
* Differential ADC with integrated amplifier
* Hardware I²C support

Thus the project can't continue until I've found a more powerful
MCU that has an acceptable toolchain and a datasheet that is not
an utter nightmare (yes, PICs fall into that category).